Structural update
Added Classes and now the pipeline works in a OOP manner. Still have a few bugs regarding connections and concurrent inserts in the database. Will fix them next. Removed the cities locations .zip folder and removed the old structure weather-aggregator.py and substituted it with the Pipeline.py file (includes Pipeline Loader and Extractor classes), and the main.py(executes the pipeline)