Refactored code and rearranged structure
Refactored some code in order to be more in accordance with OOP Standards.
Turned the Loader class into a  Thread-Safe singleton class, separated files into separated files and fixed the threading issue, as well as updated a table that wasn't with the correct length.
Added support for Ctrl-C program interruption